refactor: 치수 항목 선택 컴포넌트 분리
