+++ apps/admin/src/app/(admin)/measurement-rules/_components/measurement-rule-form/measurement-rule-form.tsx
@@ -1,18 +1,11 @@
 "use client";
 
-import { DevTool } from "@hookform/devtools";
-import { useQuery } from "@tanstack/react-query";
 import { Info, CheckSquare } from "lucide-react";
 import { PlusCircle } from "lucide-react";
-import { useState, useEffect, ComponentProps } from "react";
-import {
-  useController,
-  useForm,
-  useFormContext,
-  useWatch,
-} from "react-hook-form";
+import { useForm } from "react-hook-form";
 
 import { MeasurementRuleDefaultSection } from "./default-section";
+import { MeasurementRuleSelectSection } from "./rule-select-section";
 
 import { BasicAlert } from "@/components/Alert";
 import { Button } from "@/components/ui/button";
@@ -23,40 +16,16 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import { Checkbox } from "@/components/ui/checkbox";
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Separator } from "@/components/ui/separator";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { Form } from "@/components/ui/form";
 import {
   type MeasurementRule,
-  SLEEVE_TYPES,
   categories,
   getCategoryById,
   type SleeveType,
   isDuplicateMeasurementRule,
-  CATEGORY_ID,
-  NECKLINE_TYPES,
   NecklineType,
 } from "@/lib/data";
 import { QueryDevTools } from "@/lib/react-query";
-import { measurementRuleQueries } from "@/queries/measurement-rule";
-import { GetMeasurementRuleItemCodeResponse } from "@/services/measurement-rule";
 
 interface MeasurementRuleFormProps {
   rule?: MeasurementRule;
@@ -187,7 +156,7 @@ export function MeasurementRuleForm({
         <Card>
           <CardHeader className="pb-2">
             <div className="flex items-center justify-between">
-              <div>
+              <div className="space-y-2">
                 <CardTitle>치수 항목 선택</CardTitle>
                 <CardDescription>
                   이 규칙에 필요한 치수 항목을 선택해주세요.
@@ -210,7 +179,7 @@ export function MeasurementRuleForm({
               존재합니다. 다른 조합을 선택해주세요.
             </BasicAlert>
 
-            <MeasurementRuleSelectForm />
+            <MeasurementRuleSelectSection />
           </CardContent>
         </Card>
 
@@ -250,201 +219,4 @@ export function MeasurementRuleForm({
       <QueryDevTools control={form.control} />
     </Form>
   );
-}
-
-function MeasurementRuleSelectForm() {
-  const { data: itemCodes } = useQuery(
-    measurementRuleQueries.getMeasurementRuleItemCodeQueryOptions()
-  );
-
-  const groupedItems = transformMeasurementItems(itemCodes ?? []);
-  const [activeTab, setActiveTab] = useState<string>("상의");
-
-  // 측정 항목을 카테고리별로 그룹화
-  const itemCategories = Object.keys(groupedItems);
-  return (
-    <FormItem>
-      <Tabs
-        defaultValue="상의"
-        value={activeTab}
-        onValueChange={setActiveTab}
-        className="w-full"
-      >
-        <TabsList className="mb-6 w-full justify-start gap-1 bg-muted/50 p-1 my-4">
-          {itemCategories.map((category) => (
-            <TabsTrigger
-              key={category}
-              value={category}
-              className="px-4 py-1.5"
-            >
-              {category}
-            </TabsTrigger>
-          ))}
-        </TabsList>
-
-        {itemCategories.map((category) => (
-          <TabsContent key={category} value={category} className="mt-0">
-            <div className="space-y-6">
-              {Object.keys(groupedItems[category] ?? {}).map((section) => {
-                return (
-                  <RuleCheckList
-                    key={section}
-                    sectionName={section}
-                    sectionItems={groupedItems[category]?.[section] as any}
-                  />
-                );
-              })}
-            </div>
-          </TabsContent>
-        ))}
-      </Tabs>
-      <FormMessage />
-    </FormItem>
-  );
-}
-
-function RuleCheckList({
-  sectionItems,
-  sectionName,
-}: {
-  sectionItems: GetMeasurementRuleItemCodeResponse[];
-  sectionName: string;
-}) {
-  const form = useFormContext();
-  const selectedItems = useWatch({ name: "items" });
-
-  // 아이템 선택 처리
-  const handleItemChange = (itemId: string, checked: boolean) => {
-    const currentItems = form.getValues().items || [];
-    if (checked) {
-      form.setValue("items", [...currentItems, itemId]);
-    } else {
-      form.setValue(
-        "items",
-        currentItems.filter((id: string) => id !== itemId)
-      );
-    }
-  };
-
-  // 섹션 항목들이 모두 선택되었는지 확인
-  const isSectionFullySelected = () => {
-    const currentItems = form.getValues().items || [];
-    return sectionItems.every((item) => currentItems.includes(item.id));
-  };
-
-  // 섹션 항목들이 일부 선택되었는지 확인
-  const isSectionPartiallySelected = () => {
-    const currentItems = form.getValues().items || [];
-    const selectedCount = sectionItems.filter((item) =>
-      currentItems.includes(item.id)
-    ).length;
-    return selectedCount > 0 && selectedCount < sectionItems.length;
-  };
-
-  // 섹션의 모든 항목 선택/해제
-  const handleSectionSelectAll = (selected: boolean) => {
-    const currentItems = form.getValues().items || [];
-    if (selected) {
-      // 섹션 항목 모두 추가 (중복 제거)
-      const sectionItemIds = sectionItems.map((item) => item.id);
-      form.setValue("items", [
-        ...new Set([...currentItems, ...sectionItemIds]),
-      ]);
-    } else {
-      // 섹션 항목 모두 제거
-      const sectionItemIds = sectionItems.map((item) => item.id);
-      form.setValue(
-        "items",
-        currentItems.filter((id: string) => !sectionItemIds.includes(id))
-      );
-    }
-  };
-
-  const isFullySelected = isSectionFullySelected();
-  const isPartiallySelected = isSectionPartiallySelected();
-
-  return (
-    <div className="space-y-3">
-      <div className="flex items-center gap-2">
-        <Checkbox
-          id={`section-${sectionName}`}
-          checked={isFullySelected}
-          className={isPartiallySelected ? "opacity-70" : ""}
-          onCheckedChange={(checked) => handleSectionSelectAll(!!checked)}
-        />
-        <label
-          htmlFor={`section-${sectionName}`}
-          className="font-semibold text-gray-700"
-        >
-          {sectionName}
-        </label>
-      </div>
-      <Separator className="my-2" />
-      <div className="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-x-6 gap-y-3 pl-4">
-        {sectionItems?.map((item) => (
-          <RuleCheckItem
-            key={item.id}
-            checked={selectedItems.includes(item.id)}
-            item={item}
-            onClick={(checked) => handleItemChange(item.id, checked)}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-function RuleCheckItem({
-  checked,
-  item,
-  onClick,
-}: {
-  checked: boolean;
-  item: GetMeasurementRuleItemCodeResponse;
-  onClick: (checked: boolean) => void;
-}) {
-  return (
-    <div className="flex items-center space-x-2 ">
-      <Checkbox
-        id={`item-${item.id}`}
-        checked={checked}
-        onCheckedChange={(checked) => onClick(!!checked)}
-        className="mt-0.5"
-      />
-      <div>
-        <label
-          htmlFor={`item-${item.id}`}
-          className="font-medium leading-none cursor-pointer"
-        >
-          {item.label}
-        </label>
-      </div>
-    </div>
-  );
-}
-
-interface TransformedMeasurementItems {
-  [category: string]: {
-    [section: string]: GetMeasurementRuleItemCodeResponse[];
-  };
-}
-
-function transformMeasurementItems(
-  items: GetMeasurementRuleItemCodeResponse[]
-): TransformedMeasurementItems {
-  return items.reduce<TransformedMeasurementItems>((acc, item) => {
-    const { category, section } = item;
-
-    if (!acc[category]) {
-      acc[category] = {};
-    }
-
-    if (!acc[category][section]) {
-      acc[category][section] = [];
-    }
-
-    acc[category][section].push(item);
-
-    return acc;
-  }, {});
 }
